fix workflow tests
